requirements.txt
Pillow
torch
torchvision
pyarrow
pandas
numpy
scipy
tqdm
timm
pytorch_metric_learning
